# IvanApi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IvanApi Service",
        "title": "IvanApi Swagger",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "get": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseResult"
                        }
                    }
                }
            }
        },
        "/ops/broadcast": {
            "put": {
                "description": "update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "修改",
                "parameters": [
                    {
                        "description": "入参",
                        "name": "broadInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BroadCastUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseResult"
                        }
                    }
                }
            },
            "post": {
                "description": "update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "新增",
                "parameters": [
                    {
                        "description": "信息",
                        "name": "broadInput",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BroadCastUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseResult"
                        }
                    }
                }
            }
        },
        "/ops/getbroadcast": {
            "get": {
                "description": "get users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "获取广播",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "游戏Id",
                        "name": "gameId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "游戏版本",
                        "name": "gameVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pageIndex",
                        "name": "pageIndex",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PageResult"
                        }
                    }
                }
            }
        },
        "/test": {
            "get": {
                "description": "test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "测试",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证头",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test/second": {
            "get": {
                "description": "test2",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "测试2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BroadCastUpdateInput": {
            "type": "object",
            "properties": {
                "broadcastPosition": {
                    "type": "string"
                },
                "broadcastText": {
                    "type": "string"
                },
                "channelId": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "gameId": {
                    "type": "integer"
                },
                "gameVersion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "partnerId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "worldId": {
                    "type": "string"
                }
            }
        },
        "model.PageResult": {
            "type": "object",
            "properties": {
                "pageIndex": {
                    "description": "页签",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "页码",
                    "type": "integer"
                },
                "result": {
                    "description": "结果"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
        "model.ResponseResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态",
                    "type": "integer"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                },
                "result": {
                    "description": "结果"
                }
            }
        }
    }
}